Refactoring: added routes directory and moved user and chat routes there
