+++ main.py
@@ -1,8 +1,8 @@
 from fastapi import FastAPI
 
 from .backend.database import get_engine, Base
-from .backend.logic import (
-    user_routes, chat_routes,
+from .backend.logic.routes import (
+    chats, users,
 )
 from .tests import app
 
@@ -17,5 +17,5 @@ def startup():
     print(Base.metadata.tables)
     Base.metadata.create_all(bind=get_engine())
 
-app.include_router(user_routes.router)
-app.include_router(chat_routes.router)
+app.include_router(users.router)
+app.include_router(chats.router)
